% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorCalculation.R
\name{errorCalculation}
\alias{errorCalculation}
\title{errorCalculation}
\usage{
errorCalculation(Old, New, Transition)
}
\arguments{
\item{Old}{m by n matrix of values for n choices in m districts}

\item{New}{m by p matrix of values for n choices in m districts}

\item{Transition}{n by b matrix of transition values, all should be positive,
row sums should be one}
}
\description{
This helper function calculates the error, defined as the sum of squared
differences between the modelled results and the actual results
}

