% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorGradient.R
\name{errorGradient}
\alias{errorGradient}
\title{errorGradient}
\usage{
errorGradient(Old, New, Transition, n, p, epsilon = 1e-07)
}
\arguments{
\item{Old}{m by n matrix of values for n choices in m districts}

\item{New}{m by p matrix of values for n choices in m districts}

\item{Transition}{n by b matrix of transition values, all should be positive,
row sums should be one}

\item{n}{the number of outcomes in Old and the number of rows in Transition, provided for performance reasons}

\item{p}{the number of outcomes in New and the number of rows in Transition, provided for performance reasons}

\item{epsilon}{\eqn{\epsilon}-value for numerical approximation of gradient, defaults to 0.0000001}
}
\description{
This helper function calculates the partial gradient numerically for each element of Transition
}

