% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LTW_2011}
\alias{LTW_2011}
\title{Vote counts of the 2011 state election in Baden-Württemberg}
\format{a matrix with 70 rows and 8 columns
\describe{
\item{CDU_2011}{Vote counts for CDU}
\item{GRUENE_2011}{Vote counts for Die Grünen}
\item{SPD_2011}{Vote counts for SPD}
\item{FDP_2011}{Vote counts for FDP}
\item{LINKE_2011}{Vote counts for Die Linke}
\item{ANDERE_2011}{Sum of vote counts for all other parties}
\item{NW_2011}{number of people eligible to vote who have not cast a valid ballot}
\item{SALDO_2011_2016}{increase of number of eligible voters from 2011 to 2016}
}}
\source{
\url{http://www.statistik-bw.de/Service/Veroeff/Statistische_Analysen/803316001.pdf}
}
\usage{
LTW_2011
}
\description{
A dataset containing the vote counts of the 2011 state election in Baden-Württemberg by district
}
\keyword{datasets}

