% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LTW_2016}
\alias{LTW_2016}
\title{Vote counts of the 2016 state election in Baden-Württemberg}
\format{a matrix with 70 rows and 8 columns:
\describe{
\item{CDU_2016}{Vote counts for CDU}
\item{GRUENE_2016}{Vote counts for Die Grünen}
\item{SPD_2016}{Vote counts for SPD}
\item{FDP_2016}{Vote counts for FDP}
\item{LINKE_2016}{Vote counts for Die Linke}
\item{AFD_2016}{Vote counts for AfD}
\item{ANDERE_2016}{Sum of vote counts for all other parties}
\item{NW_2016}{number of people eligible to vote who have not cast a valid ballot}
}}
\source{
\url{http://www.statistik-bw.de/Service/Veroeff/Statistische_Analysen/803316001.pdf}
}
\usage{
LTW_2016
}
\description{
A dataset containing the vote counts of the 2016 state election in Baden-Württemberg by district
}
\keyword{datasets}

