% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimTransition.R
\name{optimTransition}
\alias{optimTransition}
\title{optimTransition}
\usage{
optimTransition(Old, New, Transition = NA, iterations = 1000,
  epsilon = 1e-07, alpha = NA, documentation = FALSE)
}
\arguments{
\item{Old}{m by n matrix of values for n choices in m districts}

\item{New}{m by p matrix of values for n choices in m districts}

\item{Transition}{n by b matrix of transition values, all should be positive,
row sums should be one, defaults to a matrix of equal values}

\item{iterations}{number of iterations, defaults to 1000}

\item{epsilon}{'\eqn{\epsilon}'-value for numerical approximation of gradient, defaults to 0.0000001}

\item{alpha}{step size \eqn{\alpha}, defaults to .01 / initial error}

\item{documentation}{boolean indicating whether long results (including
parameters of function call) shall be returned. Defaults to FALSE}
}
\description{
Core function of the package, uses gradient descent to find optimal
Transition matrix
}
\examples{
 data(LTW_2011)
 data(LTW_2016)
 result <- optimTransition (LTW_2011, LTW_2016)
 plot (result$Errors)
 View (result$Transition)
}
\seealso{
\code{\link{optimTransitionCluster}} for analysis by clusters
}

