% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimTransitionCluster.R
\name{optimTransitionCluster}
\alias{optimTransitionCluster}
\title{optimTransitionCluster}
\usage{
optimTransitionCluster(Old, New, Transition = NA, iterations = 1000,
  epsilon = 1e-07, alpha = NA, documentation = FALSE, clusterids)
}
\arguments{
\item{Old}{m by n matrix of values for n choices in m districts}

\item{New}{m by p matrix of values for n choices in m districts}

\item{Transition}{n by b matrix of transition values, all should be positive,
row sums should be one}

\item{epsilon}{'\eqn{\epsilon}'-value for numerical approximation of gradient, defaults to 0.0000001}

\item{alpha}{step size \eqn{\alpha}, defaults to .01 / initial error}
}
\description{
Splits the data set according to clusterids and calculates Transition matrices for each cluster individually
}
\seealso{
\code{\link{optimTransition}}
}

