% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Transition}
\alias{Transition}
\title{Naive initial transition matrix}
\format{a matrix with 8 rows and 8 columns}
\usage{
Transition
}
\description{
A transition matrix assuming that voters in 2016 choose the same party as in 2011
}
\keyword{datasets}

