% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionAbsolute.R
\name{transitionAbsolute}
\alias{transitionAbsolute}
\title{transitionAbsolute}
\usage{
transitionAbsolute(Old, Transition)
}
\arguments{
\item{Old}{m by n matrix of values for n choices in m districts}

\item{Transition}{n by b matrix of transition values, all should be positive,
row sums should be one}

\item{New}{m by p matrix of values for n choices in m districts}
}
\description{
Calculates the absolute values of the voter transitions, i.e. how many voters have actually shifted from one election to the next
}

