% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMatrix.R
\name{updateMatrix}
\alias{updateMatrix}
\title{updateMatrix}
\usage{
updateMatrix(Old, New, Transition, n, p, epsilon = 1e-07, alpha = 1e-12)
}
\arguments{
\item{Old}{m by n matrix of values for n choices in m districts}

\item{New}{m by p matrix of values for n choices in m districts}

\item{Transition}{n by b matrix of transition values, all should be positive,
row sums should be one}

\item{n}{the number of outcomes in Old and the number of rows in Transition, provided for performance reasons}

\item{p}{the number of outcomes in New and the number of rows in Transition, provided for performance reasons}

\item{epsilon}{'\eqn{\epsilon}'-value for numerical approximation of gradient, defaults to 0.0000001}

\item{alpha}{step size \eqn{\alpha}, defaults to 0.000000000001}
}
\description{
Updates the Transition matrix by subtracting a fraction of the gradient
}

