% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimTransition.R
\name{optimTransition}
\alias{optimTransition}
\title{optimTransition}
\usage{
optimTransition(Old, New, Transition, iterations = 1000, epsilon = 1e-07,
  alpha = 1e-12, documentation = FALSE)
}
\arguments{
\item{Old}{m by n matrix of values for n choices in m districts}

\item{New}{m by p matrix of values for n choices in m districts}

\item{Transition}{n by b matrix of transition values, all should be positive,
row sums should be one}

\item{iterations}{number of iterations, defaults to 1000}

\item{epsilon}{\eqn{\epsilon}-value for numerical approximation of gradient, defaults to 0.0000001}

\item{alpha}{step size \eqn{\alpha}, defaults to 0.000000000001}

\item{documentation}{boolean indicating whether long results (including
parameters of function call) shall be returned. Defaults to FALSE}
}
\description{
Core function of the package, uses gradient descent to find optimal
Transition matrix
}

